redis and celery added
