+++ api/requirements.txt
@@ -1,7 +1,17 @@
+amqp==5.1.1
 annotated-types==0.5.0
 anyio==3.7.1
+appdirs==1.4.4
+async-timeout==4.0.2
+attrs==23.1.0
+backports.zoneinfo==0.2.1
 bcrypt==4.0.1
+billiard==4.1.0
+celery==5.3.1
 click==8.1.4
+click-didyoumean==0.3.0
+click-plugins==1.1.1
+click-repl==0.3.0
 dnspython==2.3.0
 ecdsa==0.18.0
 email-validator==2.0.0.post2
@@ -9,20 +19,31 @@ exceptiongroup==1.1.2
 fastapi==0.100.0
 h11==0.14.0
 idna==3.4
+kombu==5.3.1
+mypy-extensions==1.0.0
+oitnb==0.2.2
 passlib==1.7.4
-pika==1.3.2
+pathspec==0.11.1
+prompt-toolkit==3.0.39
 psycopg2==2.9.6
 pyasn1==0.5.0
-python-multipart==0.0.6
 pydantic==2.0.2
 pydantic-core==2.1.2
 pydantic-settings==2.0.1
+python-dateutil==2.8.2
 python-dotenv==1.0.0
 python-jose==3.3.0
+python-multipart==0.0.6
+redis==4.6.0
+regex==2023.6.3
 rsa==4.9
 six==1.16.0
 sniffio==1.3.0
 SQLAlchemy==2.0.18
 starlette==0.27.0
+typed-ast==1.5.5
 typing-extensions==4.7.1
+tzdata==2023.3
 uvicorn==0.22.0
+vine==5.0.0
+wcwidth==0.2.6
