+++ env.sample
@@ -1,7 +1,6 @@
 API_PORT=3100
-MAIL_SERVICE_PORT=3101
 
-DATABASE_HOSTNAME=crypto_jet_db
+DATABASE_HOSTNAME=
 DATABASE_PORT=5432
 DATABASE_PASSWORD=
 DATABASE_NAME=
@@ -11,8 +10,10 @@ ALGORITHM=HS256
 ACCESS_TOKEN_SECRET_KEY=
 ACCESS_TOKEN_EXPIRE_MINUTES=600
 
-RABBITMQ_HOST=rabbitmq
-RABBITMQ_PORT=5672
-WELCOME_QUEUE_NAME=welcome_message
+NGINX_PORT=
+NGINX_SERVER_NAME=
+API_CONTAINER_NAME=api
 
-jwt_cookie_name=
+REDIS_PASSWORD=
+REDIS_HOST=redis
+REDIS_PORT=6379
