--- mail_service/main.py
@@ -1,9 +1,0 @@
-from service import listen_welcome_messages
-
-
-def main():
-    listen_welcome_messages()
-
-
-if __name__ == '__main__':
-    main()
