--- mail_service/requirements.txt
@@ -1,9 +1,0 @@
-annotated-types==0.5.0
-dnspython==2.3.0
-email-validator==2.0.0.post2
-idna==3.4
-pika==1.3.2
-pydantic==2.0.2
-pydantic-core==2.1.2
-typing-extensions==4.7.1
-pydantic-settings==2.0.1
