+++ nginx/nginx.conf
@@ -1,9 +1,11 @@
 server {
-    listen ${NGINX_PORT};
+    # listen ${NGINX_PORT};
+    listen 80;
     server_name ${SERVER_NAME};
 
     location / {
-        proxy_pass http://${API_CONTAINER_NAME}:${API_PORT};
+        # proxy_pass http://${API_CONTAINER_NAME}:${API_PORT};
+        proxy_pass http://api:3100;
         proxy_set_header Host $host;
         proxy_set_header X-Real-IP $remote_addr;
     }
